Move methods of AsciiDocLine to line
